moved things to .c file
